backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
pyTigerGraph==1.6.0
openai==1.51.0
google-generativeai==0.8.2
pydantic==2.9.2
pydantic-settings==2.5.2
